requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cffi==1.15.0
cryptography==36.0.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
whitenoise==5.3.0
